Implement useSummary hook
